backend/requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
playwright>=1.49.0
beautifulsoup4>=4.12.3
Pillow>=11.1.0
python-multipart>=0.0.19
httpx>=0.28.1
lxml>=5.3.2
